keep nice things

+++ crates/driver/src/infra/liquidity/config.rs
@@ -115,7 +115,7 @@ pub struct UniswapV3 {
     pub router: eth::ContractAddress,
 
     /// How many pools should be initialized during start up.
-    pub max_pools_to_initialize: u64,
+    pub max_pools_to_initialize: usize,
 }
 
 impl UniswapV3 {
